requirements.txt
# general  modules
django>4.2,<4.3
Pillow
djangorestframework

# env controlls
python-decouple

# third party modules
markdown       
django-filter  
drf-spectacular
djangorestframework-simplejwt
django-cors-headers
django-robots
django-debug-toolbar
requests
shapely
googlemaps

# deployment module
gunicorn
whitenoise

# refactoring and tests
black
flake8
pytest
pytest-django
Faker
pylint
pycodestyle

# background proccess and cache
celery[redis]
django-redis

# database client
psycopg[binary,pool]